Change karma to be able to run without gulp
