+++ package.json
@@ -41,6 +41,7 @@
     "jade": "~1.8.1",
     "jshint-stylish": "~1.0.0",
     "karma": "~0.12.31",
+    "karma-angular-filesort": "^0.1.0",
     "karma-jasmine": "~0.3.1",
     "karma-mocha-reporter": "^1.0.2",
     "karma-ng-html2js-preprocessor": "~0.1.2",
